Refactor the frontend to add a partial event system
